changed sunxi symbol
Changed the name SUNXI and sunxi symbols to sun8i. The sunxi symbol is
not really defined and does not include all devices which have this
kind of spinlock device. sun8i was chosen because these were the first
devices to include the hwspinlock unit. So this driver works with all
sun8i and compatible devices.