updated drivers and documentation
Move the add_action_* function to the proper place and did some more
code simplification. Also changed the device tree binding to
sun8i-a33-hwspinlock. The A33 is the first one which has the spinlock
unit mentioned in the datasheet. But it seems it already was part of
the sun6i SoCs, but not mentioned in the datasheets.