big driver and documentation update
Next attempt of mainline preparing. Changed the sun8i symbols and dt
bindings to sun6i and sun6i-a31-hwspinlock because the earliest
appearing of the same spinlock unit seems to be the Allwinner A31 SoC.
Changed the devm_add_action_* back to the more classic remove callback.
Matched the documentation to the symbol changes.